+++ autonomy/cli/helpers/chain.py
@@ -19,8 +19,9 @@
 
 """On-chain interaction helpers."""
 
+import binascii
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, cast
+from typing import Any, Dict, List, Optional, Tuple, Type, cast
 
 import click
 from aea.configurations.base import PackageConfiguration
@@ -32,7 +33,12 @@ from aea.helpers.base import IPFSHash
 from texttable import Texttable
 
 from autonomy.chain.base import ServiceState, UnitType
-from autonomy.chain.config import ChainConfigs, ChainType, ContractConfigs
+from autonomy.chain.config import (
+    ChainConfigs,
+    ChainType,
+    ContractConfig,
+    ContractConfigs,
+)
 from autonomy.chain.constants import (
     AGENT_REGISTRY_CONTRACT,
     COMPONENT_REGISTRY_CONTRACT,
@@ -85,14 +91,6 @@ try:
 except ImportError:  # pragma: nocover
     ETHEREUM_PLUGIN_INSTALLED = False
 
-try:
-    from aea_ledger_ethereum_hwi.exceptions import HWIError
-    from aea_ledger_ethereum_hwi.hwi import EthereumHWIApi
-
-    HWI_PLUGIN_INSTALLED = True
-except ImportError:  # pragma: nocover
-    HWI_PLUGIN_INSTALLED = False
-
 
 class OnChainHelper:  # pylint: disable=too-few-public-methods
     """On-chain interaction helper."""
@@ -119,7 +117,47 @@ class OnChainHelper:  # pylint: disable=too-few-public-methods
         )
 
     @staticmethod
+    def load_hwi_plugin() -> Type[LedgerApi]:  # pragma: nocover
+        """Load HWI Plugin."""
+        try:
+            from aea_ledger_ethereum_hwi.hwi import (  # pylint: disable=import-outside-toplevel
+                EthereumHWIApi,
+            )
+
+            return EthereumHWIApi
+        except ImportError as e:
+            raise click.ClickException(
+                "Hardware wallet plugin not installed, "
+                "Run `pip3 install open-aea-ledger-ethereum-hwi` to install the plugin"
+            ) from e
+        except TypeError as e:
+            raise click.ClickException(
+                'Protobuf compatibility error; Please export PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION="python" '
+                "to use the hardware wallet without any issues"
+            ) from e
+
+    @staticmethod
+    def load_crypto(
+        file: Path,
+        password: Optional[str] = None,
+    ) -> Crypto:
+        """Load crypto object."""
+        try:
+            return EthereumCrypto(
+                private_key_path=file,
+                password=password,
+            )
+        except (binascii.Error, ValueError) as e:
+            raise click.ClickException(
+                "Cannot load private key for following possible reasons\n"
+                "- Wrong key format\n"
+                "- Wrong key length\n"
+                "- Trailing spaces or new line characters"
+            ) from e
+
+    @classmethod
     def get_ledger_and_crypto_objects(
+        cls,
         chain_type: ChainType,
         key: Optional[Path] = None,
         password: Optional[str] = None,
@@ -136,13 +174,8 @@ class OnChainHelper:  # pylint: disable=too-few-public-methods
                 f"using `{ChainConfigs.get_rpc_env_var(chain_type)}` environment variable"
             )
 
-        if hwi and not HWI_PLUGIN_INSTALLED:  # pragma: nocover
-            raise click.ClickException(
-                "Hardware wallet plugin not installed, "
-                "Run `pip3 install open-aea-ledger-ethereum-hwi` to install the plugin"
-            )
-
         if hwi:
+            EthereumHWIApi = cls.load_hwi_plugin()
             identifier = EthereumHWIApi.identifier
 
         if not hwi and not ETHEREUM_PLUGIN_INSTALLED:  # pragma: nocover
@@ -154,8 +187,8 @@ class OnChainHelper:  # pylint: disable=too-few-public-methods
         if key is None:
             crypto = crypto_registry.make(identifier)
         else:
-            crypto = EthereumCrypto(
-                private_key_path=key,
+            crypto = cls.load_crypto(
+                file=key,
                 password=password,
             )
 
@@ -180,10 +213,29 @@ class OnChainHelper:  # pylint: disable=too-few-public-methods
 
         try:
             ledger_api.api.eth.default_account = crypto.address
-        except HWIError as e:  # pragma: nocover
-            raise click.ClickException(e.message)
+        except Exception as e:  # pragma: nocover
+            raise click.ClickException(str(e))
 
         return ledger_api, crypto
+
+    def check_required_enviroment_variables(
+        self, configs: Tuple[ContractConfig, ...]
+    ) -> None:
+        """Check for required enviroment variables when working with the custom chain."""
+        if self.chain_type != ChainType.CUSTOM:
+            return
+        missing = []
+        for config in configs:
+            if config.contracts[self.chain_type] is None:
+                missing.append(config)
+
+        if len(missing) == 0:
+            return
+
+        error = "Addresses for following contracts are None, please set them using their respective environment variables\n"
+        for config in missing:
+            error += f"- Set `{config.name}` address using `CUSTOM_{config.name.upper()}_ADDRESS`\n"
+        raise click.ClickException(error[:-1])
 
 
 class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
@@ -243,18 +295,27 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             contract_address = ContractConfigs.get(
                 SERVICE_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.service_registry,)
+            )
         elif self.package_type == PackageType.AGENT:
             is_service = False
             is_agent = True
             contract_address = ContractConfigs.get(
                 AGENT_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.agent_registry,)
+            )
         else:
             is_service = False
             is_agent = False
             contract_address = ContractConfigs.get(
                 COMPONENT_REGISTRY_CONTRACT.name
             ).contracts[self.chain_type]
+            self.check_required_enviroment_variables(
+                configs=(ContractConfigs.component_registry,)
+            )
 
         self.old_metadata = resolve_component_id(
             ledger_api=self.ledger_api,
@@ -357,6 +418,18 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
         component_type: UnitType = UnitType.COMPONENT,
     ) -> None:
         """Mint component."""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.registries_manager,
+                (
+                    ContractConfigs.component_registry
+                    if component_type == UnitType.COMPONENT
+                    else ContractConfigs.agent_registry
+                ),
+            )
+        )
+
         try:
             self.token_id = _mint_component(
                 ledger_api=self.ledger_api,
@@ -405,6 +478,18 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
     ) -> None:
         """Mint service"""
 
+        if self.chain_type == ChainType.CUSTOM and token is not None:
+            raise click.ClickException(
+                "Cannot use custom token for bonding on L2 chains"
+            )
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             token_id = _mint_service(
                 ledger_api=self.ledger_api,
@@ -442,6 +527,16 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
 
     def update_component(self, component_type: UnitType = UnitType.COMPONENT) -> None:
         """Update component."""
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.registries_manager,
+                (
+                    ContractConfigs.component_registry
+                    if component_type == UnitType.COMPONENT
+                    else ContractConfigs.agent_registry
+                ),
+            )
+        )
         try:
             self.token_id = _update_component(
                 ledger_api=self.ledger_api,
@@ -455,7 +550,7 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             raise click.ClickException(f"Invalid parameters provided; {e}") from e
         except ComponentMintFailed as e:
             raise click.ClickException(
-                f"Component mint failed with following error; {e}"
+                f"Component update failed with following error; {e}"
             ) from e
 
         click.echo("Component hash updated:")
@@ -481,6 +576,24 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
     ) -> None:
         """Update service"""
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
+        *_, state, _ = get_service_info(
+            ledger_api=self.ledger_api,
+            chain_type=self.chain_type,
+            token_id=cast(int, self.update_token),
+        )
+
+        if ServiceState(state) != ServiceState.PRE_REGISTRATION:
+            raise click.ClickException(
+                "Cannot update service hash, service needs to be in the pre-registration state"
+            )
+
         try:
             token_id = _update_service(
                 ledger_api=self.ledger_api,
@@ -501,7 +614,7 @@ class MintHelper(OnChainHelper):  # pylint: disable=too-many-instance-attributes
             )
         except ComponentMintFailed as e:
             raise click.ClickException(
-                f"Service mint failed with following error; {e}"
+                f"Service update failed with following error; {e}"
             ) from e
 
         click.echo("Service updated with:")
@@ -539,6 +652,11 @@ class ServiceHelper(OnChainHelper):
         token: Optional[str] = None,
     ) -> "ServiceHelper":
         """Check if service"""
+        if self.chain_type == ChainType.CUSTOM:
+            self.token = token
+            self.token_secured = False
+            return self
+
         self.token = token
         self.token_secured = is_service_token_secured(
             ledger_api=self.ledger_api,
@@ -584,6 +702,13 @@ class ServiceHelper(OnChainHelper):
                 spender=spender,
             )
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             _activate_service(
                 ledger_api=self.ledger_api,
@@ -619,6 +744,13 @@ class ServiceHelper(OnChainHelper):
                 spender=spender,
             )
 
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
+
         try:
             _register_instance(
                 ledger_api=self.ledger_api,
@@ -637,10 +769,20 @@ class ServiceHelper(OnChainHelper):
     def deploy_service(
         self,
         reuse_multisig: bool = False,
-        deployment_payload: Optional[str] = None,
+        fallback_handler: Optional[str] = None,
         timeout: Optional[float] = None,
     ) -> None:
         """Deploy a service with registration activated"""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+                ContractConfigs.gnosis_safe_proxy_factory,
+                ContractConfigs.gnosis_safe_same_address_multisig,
+                ContractConfigs.multisend,
+            )
+        )
 
         try:
             _deploy_service(
@@ -649,7 +791,7 @@ class ServiceHelper(OnChainHelper):
                 chain_type=self.chain_type,
                 service_id=self.service_id,
                 reuse_multisig=reuse_multisig,
-                deployment_payload=deployment_payload,
+                fallback_handler=fallback_handler,
                 timeout=timeout,
             )
         except ServiceDeployFailed as e:
@@ -659,6 +801,13 @@ class ServiceHelper(OnChainHelper):
 
     def terminate_service(self) -> None:
         """Terminate a service"""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
 
         try:
             _terminate_service(
@@ -674,6 +823,13 @@ class ServiceHelper(OnChainHelper):
 
     def unbond_service(self) -> None:
         """Unbond a service"""
+
+        self.check_required_enviroment_variables(
+            configs=(
+                ContractConfigs.service_manager,
+                ContractConfigs.service_registry,
+            )
+        )
 
         try:
             _unbond_service(