+++ packages/valory/protocols/tendermint/tendermint_pb2.py
@@ -4,9 +4,8 @@
 """Generated protocol buffer code."""
 from google.protobuf import descriptor as _descriptor
 from google.protobuf import descriptor_pool as _descriptor_pool
-from google.protobuf import message as _message
-from google.protobuf import reflection as _reflection
 from google.protobuf import symbol_database as _symbol_database
+from google.protobuf.internal import builder as _builder
 
 
 # @@protoc_insertion_point(imports)
@@ -18,131 +17,37 @@ DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(
     b'\n\x10tendermint.proto\x12\x1c\x61\x65\x61.valory.tendermint.v0_1_0"\xad\t\n\x11TendermintMessage\x12S\n\x05\x65rror\x18\x05 \x01(\x0b\x32\x42.aea.valory.tendermint.v0_1_0.TendermintMessage.Error_PerformativeH\x00\x12\x61\n\x0cgenesis_info\x18\x06 \x01(\x0b\x32I.aea.valory.tendermint.v0_1_0.TendermintMessage.Genesis_Info_PerformativeH\x00\x12i\n\x10get_genesis_info\x18\x07 \x01(\x0b\x32M.aea.valory.tendermint.v0_1_0.TendermintMessage.Get_Genesis_Info_PerformativeH\x00\x12o\n\x13get_recovery_params\x18\x08 \x01(\x0b\x32P.aea.valory.tendermint.v0_1_0.TendermintMessage.Get_Recovery_Params_PerformativeH\x00\x12g\n\x0frecovery_params\x18\t \x01(\x0b\x32L.aea.valory.tendermint.v0_1_0.TendermintMessage.Recovery_Params_PerformativeH\x00\x1a\x8e\x01\n\tErrorCode\x12[\n\nerror_code\x18\x01 \x01(\x0e\x32G.aea.valory.tendermint.v0_1_0.TendermintMessage.ErrorCode.ErrorCodeEnum"$\n\rErrorCodeEnum\x12\x13\n\x0fINVALID_REQUEST\x10\x00\x1a\x44\n\x1dGet_Genesis_Info_Performative\x12\r\n\x05query\x18\x01 \x01(\t\x12\x14\n\x0cquery_is_set\x18\x02 \x01(\x08\x1aG\n Get_Recovery_Params_Performative\x12\r\n\x05query\x18\x01 \x01(\t\x12\x14\n\x0cquery_is_set\x18\x02 \x01(\x08\x1a)\n\x19Genesis_Info_Performative\x12\x0c\n\x04info\x18\x01 \x01(\t\x1a.\n\x1cRecovery_Params_Performative\x12\x0e\n\x06params\x18\x01 \x01(\t\x1a\x8f\x02\n\x12\x45rror_Performative\x12M\n\nerror_code\x18\x01 \x01(\x0b\x32\x39.aea.valory.tendermint.v0_1_0.TendermintMessage.ErrorCode\x12\x11\n\terror_msg\x18\x02 \x01(\t\x12\x65\n\nerror_data\x18\x03 \x03(\x0b\x32Q.aea.valory.tendermint.v0_1_0.TendermintMessage.Error_Performative.ErrorDataEntry\x1a\x30\n\x0e\x45rrorDataEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\r\n\x05value\x18\x02 \x01(\t:\x02\x38\x01\x42\x0e\n\x0cperformativeb\x06proto3'
 )
 
-
-_TENDERMINTMESSAGE = DESCRIPTOR.message_types_by_name["TendermintMessage"]
-_TENDERMINTMESSAGE_ERRORCODE = _TENDERMINTMESSAGE.nested_types_by_name["ErrorCode"]
-_TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE = (
-    _TENDERMINTMESSAGE.nested_types_by_name["Get_Genesis_Info_Performative"]
-)
-_TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE = (
-    _TENDERMINTMESSAGE.nested_types_by_name["Get_Recovery_Params_Performative"]
-)
-_TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE = _TENDERMINTMESSAGE.nested_types_by_name[
-    "Genesis_Info_Performative"
-]
-_TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE = (
-    _TENDERMINTMESSAGE.nested_types_by_name["Recovery_Params_Performative"]
-)
-_TENDERMINTMESSAGE_ERROR_PERFORMATIVE = _TENDERMINTMESSAGE.nested_types_by_name[
-    "Error_Performative"
-]
-_TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY = (
-    _TENDERMINTMESSAGE_ERROR_PERFORMATIVE.nested_types_by_name["ErrorDataEntry"]
-)
-_TENDERMINTMESSAGE_ERRORCODE_ERRORCODEENUM = (
-    _TENDERMINTMESSAGE_ERRORCODE.enum_types_by_name["ErrorCodeEnum"]
-)
-TendermintMessage = _reflection.GeneratedProtocolMessageType(
-    "TendermintMessage",
-    (_message.Message,),
-    {
-        "ErrorCode": _reflection.GeneratedProtocolMessageType(
-            "ErrorCode",
-            (_message.Message,),
-            {
-                "DESCRIPTOR": _TENDERMINTMESSAGE_ERRORCODE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.ErrorCode)
-            },
-        ),
-        "Get_Genesis_Info_Performative": _reflection.GeneratedProtocolMessageType(
-            "Get_Genesis_Info_Performative",
-            (_message.Message,),
-            {
-                "DESCRIPTOR": _TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Get_Genesis_Info_Performative)
-            },
-        ),
-        "Get_Recovery_Params_Performative": _reflection.GeneratedProtocolMessageType(
-            "Get_Recovery_Params_Performative",
-            (_message.Message,),
-            {
-                "DESCRIPTOR": _TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Get_Recovery_Params_Performative)
-            },
-        ),
-        "Genesis_Info_Performative": _reflection.GeneratedProtocolMessageType(
-            "Genesis_Info_Performative",
-            (_message.Message,),
-            {
-                "DESCRIPTOR": _TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Genesis_Info_Performative)
-            },
-        ),
-        "Recovery_Params_Performative": _reflection.GeneratedProtocolMessageType(
-            "Recovery_Params_Performative",
-            (_message.Message,),
-            {
-                "DESCRIPTOR": _TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Recovery_Params_Performative)
-            },
-        ),
-        "Error_Performative": _reflection.GeneratedProtocolMessageType(
-            "Error_Performative",
-            (_message.Message,),
-            {
-                "ErrorDataEntry": _reflection.GeneratedProtocolMessageType(
-                    "ErrorDataEntry",
-                    (_message.Message,),
-                    {
-                        "DESCRIPTOR": _TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY,
-                        "__module__": "tendermint_pb2"
-                        # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Error_Performative.ErrorDataEntry)
-                    },
-                ),
-                "DESCRIPTOR": _TENDERMINTMESSAGE_ERROR_PERFORMATIVE,
-                "__module__": "tendermint_pb2"
-                # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage.Error_Performative)
-            },
-        ),
-        "DESCRIPTOR": _TENDERMINTMESSAGE,
-        "__module__": "tendermint_pb2"
-        # @@protoc_insertion_point(class_scope:aea.valory.tendermint.v0_1_0.TendermintMessage)
-    },
-)
-_sym_db.RegisterMessage(TendermintMessage)
-_sym_db.RegisterMessage(TendermintMessage.ErrorCode)
-_sym_db.RegisterMessage(TendermintMessage.Get_Genesis_Info_Performative)
-_sym_db.RegisterMessage(TendermintMessage.Get_Recovery_Params_Performative)
-_sym_db.RegisterMessage(TendermintMessage.Genesis_Info_Performative)
-_sym_db.RegisterMessage(TendermintMessage.Recovery_Params_Performative)
-_sym_db.RegisterMessage(TendermintMessage.Error_Performative)
-_sym_db.RegisterMessage(TendermintMessage.Error_Performative.ErrorDataEntry)
-
+_globals = globals()
+_builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
+_builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, "tendermint_pb2", _globals)
 if _descriptor._USE_C_DESCRIPTORS == False:
     DESCRIPTOR._options = None
     _TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY._options = None
     _TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY._serialized_options = b"8\001"
-    _TENDERMINTMESSAGE._serialized_start = 51
-    _TENDERMINTMESSAGE._serialized_end = 1248
-    _TENDERMINTMESSAGE_ERRORCODE._serialized_start = 582
-    _TENDERMINTMESSAGE_ERRORCODE._serialized_end = 724
-    _TENDERMINTMESSAGE_ERRORCODE_ERRORCODEENUM._serialized_start = 688
-    _TENDERMINTMESSAGE_ERRORCODE_ERRORCODEENUM._serialized_end = 724
-    _TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE._serialized_start = 726
-    _TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE._serialized_end = 794
-    _TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE._serialized_start = 796
-    _TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE._serialized_end = 867
-    _TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE._serialized_start = 869
-    _TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE._serialized_end = 910
-    _TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE._serialized_start = 912
-    _TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE._serialized_end = 958
-    _TENDERMINTMESSAGE_ERROR_PERFORMATIVE._serialized_start = 961
-    _TENDERMINTMESSAGE_ERROR_PERFORMATIVE._serialized_end = 1232
-    _TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY._serialized_start = 1184
-    _TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY._serialized_end = 1232
+    _globals["_TENDERMINTMESSAGE"]._serialized_start = 51
+    _globals["_TENDERMINTMESSAGE"]._serialized_end = 1248
+    _globals["_TENDERMINTMESSAGE_ERRORCODE"]._serialized_start = 582
+    _globals["_TENDERMINTMESSAGE_ERRORCODE"]._serialized_end = 724
+    _globals["_TENDERMINTMESSAGE_ERRORCODE_ERRORCODEENUM"]._serialized_start = 688
+    _globals["_TENDERMINTMESSAGE_ERRORCODE_ERRORCODEENUM"]._serialized_end = 724
+    _globals["_TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE"]._serialized_start = 726
+    _globals["_TENDERMINTMESSAGE_GET_GENESIS_INFO_PERFORMATIVE"]._serialized_end = 794
+    _globals[
+        "_TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE"
+    ]._serialized_start = 796
+    _globals[
+        "_TENDERMINTMESSAGE_GET_RECOVERY_PARAMS_PERFORMATIVE"
+    ]._serialized_end = 867
+    _globals["_TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE"]._serialized_start = 869
+    _globals["_TENDERMINTMESSAGE_GENESIS_INFO_PERFORMATIVE"]._serialized_end = 910
+    _globals["_TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE"]._serialized_start = 912
+    _globals["_TENDERMINTMESSAGE_RECOVERY_PARAMS_PERFORMATIVE"]._serialized_end = 958
+    _globals["_TENDERMINTMESSAGE_ERROR_PERFORMATIVE"]._serialized_start = 961
+    _globals["_TENDERMINTMESSAGE_ERROR_PERFORMATIVE"]._serialized_end = 1232
+    _globals[
+        "_TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY"
+    ]._serialized_start = 1184
+    _globals[
+        "_TENDERMINTMESSAGE_ERROR_PERFORMATIVE_ERRORDATAENTRY"
+    ]._serialized_end = 1232
 # @@protoc_insertion_point(module_scope)