+++ tests/test_autonomy/test_cli/test_helpers/test_chain_helpers.py
@@ -19,12 +19,13 @@
 
 """Test chain helpers."""
 
+import tempfile
+from pathlib import Path
 from unittest import mock
 
 import click
 import pytest
 from aea.configurations.data_types import PackageType
-from aea_ledger_ethereum_hwi.hwi import EthereumHWIApi, EthereumHWICrypto
 from aea_test_autonomy.configurations import ETHEREUM_KEY_DEPLOYER
 
 from autonomy.chain.base import ServiceState
@@ -52,6 +53,14 @@ _ = registry_contracts.service_manager
 _ = registry_contracts.registries_manager
 _ = registry_contracts.component_registry
 _ = registry_contracts.service_manager
+
+
+def _get_ledger_and_crypto_objects_patch() -> mock._patch:
+    return mock.patch.object(
+        OnChainHelper,
+        "get_ledger_and_crypto_objects",
+        return_value=(mock.MagicMock(), mock.MagicMock()),
+    )
 
 
 class TestMintComponentMethod:
@@ -124,6 +133,128 @@ class TestMintComponentMethod:
                 ).verify_nft().verify_component_dependencies(
                     dependencies=(),  # type: ignore
                 ).publish_metadata().mint_component()
+
+
+class TestRequiredEnvVars:
+    """Test required env var check works."""
+
+    _component_failure = (
+        "Addresses for following contracts are None, please set them using their respective environment variables\n"
+        "- Set `registries_manager` address using `CUSTOM_REGISTRIES_MANAGER_ADDRESS`\n"
+        "- Set `component_registry` address using `CUSTOM_COMPONENT_REGISTRY_ADDRESS`"
+    )
+
+    _service_failure = (
+        "Addresses for following contracts are None, please set them using their respective environment variables\n"
+        "- Set `service_manager` address using `CUSTOM_SERVICE_MANAGER_ADDRESS`\n"
+        "- Set `service_registry` address using `CUSTOM_SERVICE_REGISTRY_ADDRESS`"
+    )
+
+    def test_component_mint(self) -> None:
+        """Test component mint env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._component_failure,
+        ):
+            mint_helper = MintHelper(
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.mint_component()
+
+    def test_component_update(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._component_failure,
+        ):
+            mint_helper = MintHelper(
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.update_component()
+
+    def test_servic_mint(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._service_failure,
+        ):
+            mint_helper = MintHelper(
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.mint_service(1, 1, 1)
+
+    def test_service_update(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._service_failure,
+        ):
+            mint_helper = MintHelper(
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.update_service(1, 1, 1)
+
+    def test_activate_service(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._service_failure,
+        ):
+            mint_helper = ServiceHelper(
+                service_id=1,
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.check_is_service_token_secured().activate_service()
+
+    def test_register_instances(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._service_failure,
+        ):
+            mint_helper = ServiceHelper(
+                service_id=1,
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.check_is_service_token_secured().register_instance(["0x"], [1])
+
+    def test_deploy_instances(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=(
+                "Addresses for following contracts are None, please set them using their respective environment variables\n"
+                "- Set `service_manager` address using `CUSTOM_SERVICE_MANAGER_ADDRESS`\n"
+                "- Set `service_registry` address using `CUSTOM_SERVICE_REGISTRY_ADDRESS`\n"
+                "- Set `gnosis_safe_proxy_factory` address using `CUSTOM_GNOSIS_SAFE_PROXY_FACTORY_ADDRESS`\n"
+                "- Set `gnosis_safe_same_address_multisig` address using `CUSTOM_GNOSIS_SAFE_SAME_ADDRESS_MULTISIG_ADDRESS`"
+            ),
+        ):
+            mint_helper = ServiceHelper(
+                service_id=1,
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.check_is_service_token_secured().deploy_service()
+
+    def test_unbond(self) -> None:
+        """Test component update env vars."""
+        with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+            click.ClickException,
+            match=self._service_failure,
+        ):
+            mint_helper = ServiceHelper(
+                service_id=1,
+                chain_type=ChainType.CUSTOM,
+                key=ETHEREUM_KEY_DEPLOYER,
+            )
+            mint_helper.check_is_service_token_secured().unbond_service()
 
 
 @pytest.mark.parametrize(
@@ -224,8 +355,10 @@ def test_unbond_service_contract_failure() -> None:
             ).unbond_service()
 
 
+@pytest.mark.skip(reason="https://github.com/valory-xyz/open-aea/issues/671")
 def test_get_ledger_and_crypto_objects() -> None:
     """Test `get_ledger_and_crypto_objects` for hardware wallet support"""
+    from aea_ledger_ethereum_hwi.hwi import EthereumHWIApi, EthereumHWICrypto
 
     with mock.patch.object(EthereumHWICrypto, "entity"):
         ledger_api, crypto = OnChainHelper.get_ledger_and_crypto_objects(
@@ -306,3 +439,53 @@ def test_verify_service_dependencies_failures() -> None:
         ).verify_service_dependencies(
             agent_id=1
         )
+
+
+def test_token_secure_check_on_custom_chain() -> None:
+    """Test token_secure is False on custom chain."""
+
+    with _get_ledger_and_crypto_objects_patch():
+        service_helper = ServiceHelper(
+            service_id=1, chain_type=ChainType.CUSTOM, key=ETHEREUM_KEY_DEPLOYER
+        )
+        assert service_helper.check_is_service_token_secured().token_secured is False
+
+
+def test_mint_with_token_on_custom_chain() -> None:
+    """Test minting with token on L2 chains fail."""
+
+    with _get_ledger_and_crypto_objects_patch(), pytest.raises(
+        click.ClickException, match="Cannot use custom token for bonding on L2 chains"
+    ):
+        MintHelper(  # nosec
+            chain_type=ChainType.CUSTOM, key=ETHEREUM_KEY_DEPLOYER
+        ).mint_service(
+            number_of_slots=1,
+            cost_of_bond=1,
+            threshold=1,
+            token="0x",
+        )
+
+
+@pytest.mark.parametrize(
+    argnames="key",
+    argvalues=(
+        "0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e\n",
+        "df57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656",
+        "0x9fe46736679d2d9a65f0992f2272de9f3c7fa6e0",
+    ),
+)
+def test_wrong_private_key_format(key: str) -> None:
+    """Test an error is raised if the private key format is wrong."""
+    with tempfile.TemporaryDirectory() as temp_dir, pytest.raises(
+        click.ClickException,
+        match=(
+            "Cannot load private key for following possible reasons\n"
+            "- Wrong key format\n"
+            "- Wrong key length\n"
+            "- Trailing spaces or new line characters"
+        ),
+    ):
+        file = Path(temp_dir, "key.txt")
+        file.write_text(key)
+        OnChainHelper.load_crypto(file=file)