+++ tests/test_autonomy/test_cli/test_service/test_service_manager.py
@@ -23,6 +23,7 @@ import binascii
 from typing import List, Optional
 from unittest import mock
 
+import pytest
 from aea.components.base import load_aea_package
 from aea.configurations.data_types import PackageType
 from aea.configurations.loader import load_configuration_object
@@ -117,7 +118,6 @@ class BaseServiceManagerTest(BaseChainInteractionTest):
     def deploy_service(
         self,
         service_id: int,
-        deployment_payload: Optional[str] = None,
         reuse_multisig: bool = False,
     ) -> None:
         """Deploy service."""
@@ -127,7 +127,6 @@ class BaseServiceManagerTest(BaseChainInteractionTest):
             crypto=self.crypto,
             chain_type=self.chain_type,
             service_id=service_id,
-            deployment_payload=deployment_payload,
             reuse_multisig=reuse_multisig,
         )
 
@@ -738,11 +737,16 @@ class TestServiceRedeploymentWithSameMultisig(BaseServiceManagerTest):
         tx_digest = self.ledger_api.send_signed_transaction(stx)
         self.ledger_api.get_transaction_receipt(tx_digest)
 
-    def test_redeploy(self) -> None:
+    @pytest.mark.parametrize(
+        argnames="n_owners",
+        argvalues=(1, 2),
+    )
+    def test_redeploy(self, n_owners: int) -> None:
         """Test redeploy service with same multisig."""
-        n = 2
-        instances = self.generate_and_fund_keys(n=n)
-        service_id = self.mint_service(number_of_slots_per_agent=n, threshold=n)
+        instances = self.generate_and_fund_keys(n=n_owners)
+        service_id = self.mint_service(
+            number_of_slots_per_agent=n_owners, threshold=n_owners
+        )
         self.activate_service(service_id=service_id)
         for instance in instances:
             self.register_instances(
@@ -758,7 +762,7 @@ class TestServiceRedeploymentWithSameMultisig(BaseServiceManagerTest):
         )
         self.remove_owners(multisig_address=multisig_address, owners=instances)
 
-        new_instances = self.generate_and_fund_keys(n=n)
+        new_instances = self.generate_and_fund_keys(n=n_owners)
         self.activate_service(service_id=service_id)
         for instance in new_instances:
             self.register_instances(